+++ Assignments/W5-Command/Receiver.cs
@@ -7,22 +7,22 @@ namespace W5_Command
 {
     public class Receiver
     {
-        public void Add(int a, int b)
+        public void Add(double a, double b)
         {
             Console.WriteLine(a + " + " + b + " = " + (a + b));;
         }
 
-        public void Substract(int a, int b)
+        public void Substract(double a, double b)
         {
             Console.WriteLine(a + " - " + b + " = " + (a - b)); ;
         }
 
-        public void Multiply(int a, int b)
+        public void Multiply(double a, double b)
         {
             Console.WriteLine(a + " * " + b + " = " + (a * b)); ;
         }
 
-        public void Divide(int a, int b)
+        public void Divide(double a, double b)
         {
             Console.WriteLine(a + " / " + b + " = " + (a / b)); ;
         }
